requirements.txt
torch
numpy
einops
torchvision
tqdm
diffusers
accelerate
transformers
Pillow
wandb
git+https://github.com/openai/CLIP.git

# for metric calcluations:
pytorch-fid
thop

#for data processing:
h5py
webdataset
img2dataset